review changes; normalize update
